Consider local project to be default for 'repo start'
The requirement to explicitly specify the local project when starting
a new repo branch is somewhat counter intuitive.

This patch uses the current directory's git tree as the default
project.

Tested by running

  'repo start <name>'

observed that the result is the same as if running

  'repo start <name> .'

Note: Merged upstream and should be in v1.12.35+:
https://gerrit.googlesource.com/git-repo/+/e5c0ea0a959e72ddf6ace9c5a75d86de9e7fc1df

Change-Id: I7d8bccb72e191f159a5287394fadad644e39634b